Add basic 4chan image crawler
Download all images from a board, the archive of a board, or a single
thread.

The crawler is currently single-threaded which, in combination with
Pythons abysmal performace, makes it dog slow. It's good enough
for a start, but to actually be usable, it needs to eventually
be rewritten in a saner language.